Move VMVX passes from `Passes.td` -> `VMVX/Passes.td`.
