+++ compiler/src/iree/compiler/Codegen/VMVX/VMVXPasses.h
@@ -44,6 +44,13 @@ std::unique_ptr<OperationPass<mlir::ModuleOp>> createVMVXLinkExecutablesPass();
 
 /// Populates passes needed to link HAL executables across VMVX targets.
 void buildVMVXLinkingPassPipeline(OpPassManager &passManager);
+
+//----------------------------------------------------------------------------//
+// Register VMVX Passes
+//----------------------------------------------------------------------------//
+
+void registerCodegenVMVXPasses();
+
 } // namespace iree_compiler
 } // namespace mlir
